requirements.txt
fastapi
matplotlib
matplotlib-inline
notebook
numpy
pandas
pandas-datareader
passlib
Pillow
pipenv
poetry
poetry-core
poetry-plugin-export
pydot
pytest
scikit-learn
scikit-surprise
scipy
seaborn
tensorboard
tensorboard-data-server
tensorflow
tensorflow-estimator
python-multipart
tensorflow-intel
tensorflow-io-gcs-filesystem
tomlkit
urllib3
uvicorn
virtualenv
virtualenv-clone
websocket-client
websockets
keras
lfs
docker